extract clipped image shapes into ShapeImage component
As a cleanup, I extracted the clipped images left and right-aligned shapes into a ShapeImage component. This implementation is using the cssblocks state feature to define which side the image should align to.
This aligns with what we do with other shapes, help prevent copy-paste mistakes, and allow changes to one location to spread to all usages of the component.

There should be no visual changes.