Make `yarn start` render limit content archive by default
Since most times when we are working on the website we do not need
the full content archive to be rendered, either because we are working
on a specific page or on a blog post that should be within the range
of the limited content archive, I think making `yarn start` is a quality
of life improvement.
I replaced `start-dev` with a `start:production` script for when
the full content archive is required.

Let me know if I am missing something in our pre-rendering pipeline
that might be affected by this change.