turn off prosewrap for markdown files
previously we have had prosewrap turned on to make sure there is consistency in the line width for markdown files. This changes the prosewrap to have a value of `never` to bring the consistency in the opposite way. This means that if you want to view wrapped in your editor it's down to your settings, it also make diffs less complex when working on blog posts